Make vote hash browser-independent
Use a more canonical form than JSON to compute the hash of a vote.

Rationale: according to JSON.stringify() documentation, properties of
non-array objects are not guaranteed to be stringified in any
particular order. Guaranteeing the same stringification by JavaScript
code and Python code is even more sketchy.

Evidence of lurking bug: hashes listed in the audited ballots page
(computed with a Python stringification) do not correspond to the ones
the user sees in the booth (when voting and in the verifier). The
hashes computed in the voting process and during verification are the
same when both are executed in the same browser, but I've observed
differences between Firefox and Chrome.

Another alternative would have been to store directly the
stringification as it is submitted by the booth, but that looks
fragile and more intrusive.