fix CheapQueue BUG, re-enable its multiproc use
CheapQueue differentiates between "master" and "worker"
by comparing the pid of the running process and
the pid of the process that created the CheapQueue object.

Saving of this initial pid was done within the global
scope of the class definition instead of its __init__(),
and therefore was only working when class and __init__()
were called within the same process, which was obviously
not true in our production django/celery deployment.