Add LICENSE.md; insert some example copyright notices.
This commit adds LICENSE.md with the text of the Apache 2.0 license.
It also adds copyright notices at the tops of a few files, to
demonstrate the way these notices should be formatted.  Note that the
copyright years in the notices refer to the year or years in which the
file has undergone significant changes.  Files whose copyright is
owned by another project (for example see autoretry.py) should not
have their copyright notices changed.

You can use copyright-divination.sh, which is available at
https://github.com/OpenTechStrategies/ots-tools/blob/master/copyright-divination.sh,
to help determine which years should be included in the copyright
notice for each file.  Feel free to use that script to automate the
process of adding copyright headers to the rest of the files that need
them, but remember that you might know things that the script can't
know.  For example, a year in which the only change to a file was that
you changed the indentation style would not be significant from a
copyright standpoint, but the script would see that every line was
changed and recommend that year be included in the notice.